renamed randomtokenstring to randomtoken
